% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mec.r
\name{bitprint}
\alias{bitprint}
\title{bitprint}
\usage{
bitprint(str)
}
\arguments{
\item{str}{A string vector.}
}
\description{
Prints the bits of a string.  Useful to see that \code{encode()}
does indeed do what the docs say it does.
}
\examples{
str <- "a b a c a b b"
bitprint(str)
}

