% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mec.r
\name{decode}
\alias{decode}
\title{decode}
\usage{
decode(str)
}
\arguments{
\item{str}{A string vector.}
}
\value{
A string vector of the same length as the input.
}
\description{
Decode a string according to the MECA (Missy Elliott Cryptographic
Algorithm).  It's the reverse process of the MECA encoding.
}
\examples{
str <- "a b a c a b b"
decode(str)
}

