% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mec.r
\name{encode}
\alias{encode}
\title{encode}
\usage{
encode(str)
}
\arguments{
\item{str}{A string vector.}
}
\value{
A string vector of the same length as the input.
}
\description{
Encode a string according to the MECA (Missy Elliott Cryptographic
Algorithm).  Namely, we shift the bits down, flip it, and reverse
it.
}
\examples{
str <- "a b a c a b b"
encode(str)
}

