% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mec-package.r
\docType{package}
\name{mec}
\alias{mec}
\alias{mec-package}
\title{Missy Elliott Cryptography}
\description{
Missy Elliott Cryptography
}
\details{
This package implements the MECA, or Missy Elliott Cryptography Algorithm.
Shift yo bits, flip it, and reverse it.

Idea from an ancient xkcd https://xkcd.com/153/.
}
\author{
Drew Schmidt \email{wrathematics@gmail.com}
}
\keyword{package}

