% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mec-package.r
\docType{package}
\name{mec}
\alias{mec}
\alias{mec-package}
\title{Missy Elliott Cryptography}
\description{
Missy Elliott Cryptography
}
\details{
Shift yo bits, flip it, and reverse it.
}
\author{
Drew Schmidt \email{wrathematics@gmail.com}
}
\keyword{package}

